Started bones-from-script
Made the basics of pulling all the bones from a JS object array. The main reason is for the tooltips but might as well create all from the data. Nicer to edit, and it doesn’t clutter the vanilla HTML as much. Moved the place for the Finishes from the middle to the left column. The idea is that once all slots have been filled, the normal bones disappear, and the mods and finishes show up (left column). Also need to add a remaining slots indicator to the middle... And make that purple! And finally, removed that weird favicon which didn’t work anyway. I’ll mess with that later.